+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `engine` used to run checks
   - `sql_server_type` to define the type of SQL Server to use when engine is `sql`
 - Changelog support for `Info` and `Terms` blocks.
+- `datacontract import` now has `--output` option for saving Data Contract to file
 
 ### Changed
 - Changelog support for custom extension keys in `Models` and `Fields` blocks.